+++ packages/client/package.json
@@ -15,6 +15,7 @@
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-firebase-hooks": "^5.1.1",
     "react-router-dom": "^6.11.0",
     "react-scripts": "5.0.1",
     "react-slick": "^0.29.0",
